content/media/video/test/test_wav_onloadedmetadata.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=467972
-->
<head>
  <title>Wav Test for Bug 467972</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=467972">Mozilla Bug 467972</a>

<video id="v"
       src="r11025_s16_c1.wav"
       onloadedmetadata="return loadedMetaData();"
       onended="playbackEnded();"
       onloadedfirstframe="return loadedFirstFrame();"       
       onseeking="seekStarted();"
       onseeked="seekEnded();"       
       controls></video>

<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 467972 **/


var gEnded = false;
var gSeekStarted = false;
var gSeekEnded = false;
var gLoadedFirstFrameCount = 0;
var gLoadedMetaDataCount = 0;

function get(id) {
  return document.getElementById(id);
}

function video() {
  return get('v');
}

function seekStarted() {
  gSeekStarted = true;
}

function seekEnded() {
  gSeekEnded = true;
  video().play();
}

function loadedFirstFrame() {
  gLoadedFirstFrameCount++;
  ok(gLoadedFirstFrameCount <= 1, "No more than 1 onloadedfirstframe events");
}

function loadedMetaData() {
  gLoadedMetaDataCount++;
  ok(gLoadedMetaDataCount <= 1, "No more than 1 onloadedmetadata events");
  video().play();
  return false;
}

function playbackEnded() {
  if (!gEnded) {
    video().currentTime = 0;
    gEnded = true;
  } else {
    ok(gSeekEnded, "Should have received seekended");
    ok(gSeekStarted, "Should have received seekstarted");
    ok(gLoadedFirstFrameCount == 1, "Should have 1 onloadedfirstframe event");
    ok(gLoadedMetaDataCount == 1, "Should have 1 onloadedmetadata event");
    SimpleTest.finish();    
  }  
  return false;
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
